requirements.txt
# Core web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.1.0

# HTML templating
jinja2==3.1.3

# Google Cloud client libraries
google-cloud-firestore==2.13.0
google-cloud-pubsub==2.21.0
google-cloud-functions==1.0.4
google-api-python-client==2.126.0

# AI/ML Components
dialogflow==1.1.0

# Testing and Development
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
requests==2.31.0

# Type checking and linting
mypy==1.8.0
black==24.1.0
flake8==7.0.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
structlog==24.1.0
